requirements.txt
requests
Pillow
typer[all]